refactor(formation): extract methods form component/normalize
