+++ runtime/formation/src/component/normalize/index.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const { events } = require('./events')
+const { extensions } = require('./extensions')
+const { operations } = require('./operations')
+
+exports.events = events
+exports.extensions = extensions
+exports.operations = operations
